refactor: [EL-4864] address PR review comments
- Replace MUI Button with BaseBtn in OpenApiDelegatedLoginButton and OpenApiOAuthStatus
- Create index.js for openapi lib/hooks and update imports
- Deduplicate credentials config in OpenApiOAuthStatus using isReferenceMode flag
