+++ src/[fsd]/features/openapi/lib/hooks/index.js
@@ -1,0 +1,2 @@
+export { useOpenApiCheckConnection } from './useOpenApiCheckConnection.hooks';
+export { useResolvedOpenApiConfig } from './useResolvedOpenApiConfig.hooks';
